requirements.txt
requests==2.28.1
python_telegram_bot==13.13
APScheduler==3.6.3
python-dotenv==0.20.0
telegram==0.0.1
tgtg==0.13.0
